create big ol' fetcher file
